Powp/tests/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
openai==0.28.1
python-dotenv==1.0.0